% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjSDM_configs.R
\name{Adagrad}
\alias{Adagrad}
\title{Adagrad}
\usage{
Adagrad(
  lr_decay = 0,
  weight_decay = 0,
  initial_accumulator_value = 0,
  eps = 1e-10
)
}
\arguments{
\item{lr_decay}{learning rate decay}

\item{weight_decay}{l2 penalty on weights}

\item{initial_accumulator_value}{initial value of the gradient accumulator}

\item{eps}{fuzz factor}
}
\value{
Anonymous function that returns optimizer when called.
}
\description{
Adaptive subgradient optimizer, see Duchi et al., 2011
}
\references{
Duchi, J., Hazan, E., & Singer, Y. (2011). Adaptive subgradient methods for online learning and stochastic optimization. Journal of Machine Learning Research, 12(7).
}
