% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjSDM_configs.R
\name{Adam}
\alias{Adam}
\title{Adam}
\usage{
Adam(betas = c(0.9, 0.999), eps = 1e-08, weight_decay = 0, amsgrad = FALSE)
}
\arguments{
\item{betas}{exponential decay rates}

\item{eps}{fuzz factor}

\item{weight_decay}{l2 penalty on weights}

\item{amsgrad}{use the AMSGrad variant}
}
\value{
Anonymous function that returns optimizer when called.
}
\description{
Adam optimizer, see Kingma and Ba, 2014
}
\references{
Kingma, D. P., & Ba, J. (2014). Adam: A method for stochastic optimization. arXiv preprint arXiv:1412.6980.
}
