% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjSDM_configs.R
\name{AdamW}
\alias{AdamW}
\title{AdamW}
\usage{
AdamW(betas = c(0.9, 0.999), eps = 1e-08, weight_decay = 0.01, amsgrad = FALSE)
}
\arguments{
\item{betas}{exponential decay rates}

\item{eps}{fuzz factor}

\item{weight_decay}{decoupled l2 penalty on weights}

\item{amsgrad}{use the AMSGrad variant}
}
\value{
Anonymous function that returns optimizer when called.
}
\description{
Adam with decoupled weight decay, see Loshchilov and Hutter, 2019
}
\references{
Loshchilov, I., & Hutter, F. (2019). Decoupled weight decay regularization. International Conference on Learning Representations.
}
