% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{add_curve}
\alias{add_curve}
\title{add_curve
curve plotting 'engine'}
\usage{
add_curve(
  p1 = NULL,
  p2 = NULL,
  n = 10,
  spar = 0.7,
  col = "black",
  species = TRUE,
  radius = 5,
  lwd = 1
)
}
\arguments{
\item{p1}{first point}

\item{p2}{second points}

\item{n}{number of points for spline}

\item{spar}{smoothing value}

\item{col}{curve's color}

\item{species}{draw species line}

\item{radius}{radius}

\item{lwd}{curve lwd}
}
\description{
add_curve
curve plotting 'engine'
}
