% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{add_legend}
\alias{add_legend}
\title{add_legend
add legend to circular plot}
\usage{
add_legend(cols = 1:11, range = c(-1, 1), radius = 5, angles = c(110, 70))
}
\arguments{
\item{cols}{colors for gradients}

\item{range}{gradient range}

\item{radius}{radius}

\item{angles}{angles, start and end values in degree}
}
\description{
add_legend
add legend to circular plot
}
