% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{add_species_arrows}
\alias{add_species_arrows}
\title{add_species_arrows
add species arrows to circle}
\usage{
add_species_arrows(radius = NULL, label = "Species", reverse = TRUE,
  start = 150, end = 270)
}
\arguments{
\item{radius}{radius}

\item{label}{label between arrows}

\item{reverse}{reverse label}

\item{start}{start point for arrow in degree}

\item{end}{end point for arrow in degree}
}
\description{
add_species_arrows
add species arrows to circle
}
