% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjSDM_configs.R
\name{bioticStruct}
\alias{bioticStruct}
\title{biotic structure}
\usage{
bioticStruct(df = NULL, lambda = 0, alpha = 0.5, on_diag = TRUE)
}
\arguments{
\item{df}{degree of freedom for covariance parametrization, if \code{NULL} df is set to \code{ncol(Y)/2}}

\item{lambda}{lambda penality}

\item{alpha}{weighting between LASSO and ridge}

\item{on_diag}{regularization on diagonals}
}
\description{
define biotic interaction structur
}
