% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkModel}
\alias{checkModel}
\title{check model
check model and rebuild if necessary}
\usage{
checkModel(object)
}
\arguments{
\item{object}{of class sjSDM}
}
\description{
check model
check model and rebuild if necessary
}
