% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjSDM.R
\name{coef.sjSDM}
\alias{coef.sjSDM}
\title{Return coefficients from a fitted sjSDM model}
\usage{
\method{coef}{sjSDM}(object, ...)
}
\arguments{
\item{...}{optional arguments for compatibility with the generic function, no function implemented}

\item{x}{a model fitted by \code{\link{sjSDM}}}
}
\description{
Return coefficients from a fitted sjSDM model
}
