% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sLVM.R
\name{coef.sLVM}
\alias{coef.sLVM}
\title{Return coefficients from a fitted sLVM model}
\usage{
\method{coef}{sLVM}(object, ...)
}
\arguments{
\item{object}{a model fitted by \code{\link{sLVM}}}

\item{...}{optional arguments for compatibility with the generic function, no function implemented}
}
\description{
Return coefficients from a fitted sLVM model
}
