% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn.R
\name{compile}
\alias{compile}
\title{Configure an object}
\usage{
compile(object, df = NULL, l1_cov = 0, l2_cov = 0, optimizer = NULL, ...)
}
\arguments{
\item{object}{a model created by \code{\link{sjSDM_model}}}

\item{df}{degree of freedom for covariance matrix}

\item{l1_cov}{lasso regularization on covariances}

\item{l2_cov}{ridge regularization on covariances}

\item{optimizer}{optimizer to use}

\item{...}{degree of freedom, l1_cov, l2_cov}
}
\description{
Finalizes or completes an object.
Return coefficients from a fitted sjSDM model
}
