% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{curve_text}
\alias{curve_text}
\title{curve_text
plot curved text}
\usage{
curve_text(pos = 0, label = "", radius = 5, reverse = FALSE,
  middle = FALSE, extend = 1.1, ...)
}
\arguments{
\item{pos}{position in degree}

\item{label}{label}

\item{radius}{radius}

\item{reverse}{in reverse order}

\item{middle}{text in the middle}

\item{extend}{extend char lengths, default 1.1}

\item{...}{passed to \code{\link{graphics::text}}}
}
\description{
curve_text
plot curved text
}
