% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjSDM_configs.R
\name{DiffGrad}
\alias{DiffGrad}
\title{DiffGrad}
\usage{
DiffGrad(betas = c(0.9, 0.999), eps = 1e-08, weight_decay = 0)
}
\arguments{
\item{betas}{betas}

\item{eps}{eps}

\item{weight_decay}{weight_decay}
}
\value{
Anonymous function that returns optimizer when called.
}
\description{
DiffGrad
}
