% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjSDM_configs.R
\name{envLinear}
\alias{envLinear}
\title{linear}
\usage{
envLinear(data = NULL, formula = NULL, lambda = 0, alpha = 0.5)
}
\arguments{
\item{data}{matrix of environmental predictors}

\item{formula}{formula object for predictors}

\item{lambda}{lambda penality}

\item{alpha}{weighting between LASSO and ridge}
}
\description{
create linear env covariates
}
