% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn.R
\name{fit}
\alias{fit}
\title{Fit sjSDM_model}
\usage{
fit(object, X = NULL, Y = NULL, batch_size = 25L, epochs = 100L, parallel = 0L)
}
\arguments{
\item{object}{a model created by \code{\link{sjSDM_model}}}

\item{X}{environmental matrix}

\item{Y}{species occurence matrix}

\item{batch_size}{batch_size}

\item{epochs}{number of epochs}

\item{parallel}{number of cores for data loader}
}
\description{
fit sjSDM_model in n epochs
}
