% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generateSpatialEV}
\alias{generateSpatialEV}
\title{Generate spatial eigenvectors}
\usage{
generateSpatialEV(coords = NULL, threshold = 0)
}
\arguments{
\item{coords}{matrix or data.frame of coordinates}

\item{threshold}{ignore distances greater than threshold}
}
\value{
Matrix of spatial eigenvectors.
}
\description{
Generates a Moran's eigenvector map of the distance matrix. See Dray, Legendre, and Peres-Neto, 2006 for more information.
}
\references{
Dray, S., Legendre, P., & Peres-Neto, P. R. (2006). Spatial modelling: a comprehensive framework for principal coordinate analysis of neighbour matrices (PCNM). Ecological modelling, 196(3-4), 483-493.
}
