% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sLVM.R
\name{getCI}
\alias{getCI}
\title{Get Credible Intervals}
\usage{
getCI(object, lower = 0.025, upper = 0.975)
}
\arguments{
\item{object}{a model fitted by \code{\link{sLVM}}}

\item{lower}{lower quantile}

\item{upper}{upper quantile}
}
\description{
Get Credible Intervals
}
