% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{getImportance}
\alias{getImportance}
\title{getImportance}
\usage{
getImportance(beta, sp = NULL, association, covX, covSP = NULL)
}
\arguments{
\item{beta}{abiotic weights}

\item{sp}{spatial weights}

\item{association}{species associations}

\item{covX}{environmental covariance matrix}

\item{covSP}{spatial covariance matrix}
}
\description{
variation partitioning with coefficients
}
\author{
Maximilian Pichler
}
