% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjSDM.R
\name{getSe}
\alias{getSe}
\title{Post hoc calculation of standard errors}
\usage{
getSe(object, step_size = NULL, parallel = 0L)
}
\arguments{
\item{object}{a model fitted by \code{\link{sjSDM}}}

\item{step_size}{batch size for stochastic gradient descent}

\item{parallel}{number of cpu cores for the data loader, only necessary for large datasets}
}
\value{
The object passed to this function but the \code{object$se} field contains the standard errors now
}
\description{
Post hoc calculation of standard errors
}
