% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getWeights}
\alias{getWeights}
\alias{getWeights.sjSDM_model}
\alias{getWeights.sjSDM_DNN}
\title{Get weights}
\usage{
getWeights(object)

\method{getWeights}{sjSDM_model}(object)

\method{getWeights}{sjSDM_DNN}(object)
}
\arguments{
\item{object}{object of class \code{\link{sjSDM}} with \code{\link{envDNN}} object or of class \code{\link{sjSDM_model}}}
}
\value{
\itemize{
 \item layers - list of layer weights
 \item sigma - weight to construct covariance matrix
}
}
\description{
return weights of each layer
}
