% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_sjSDM_py_available}
\alias{is_sjSDM_py_available}
\title{is_sjSDM_py_available
check whetcher torch is available}
\usage{
is_sjSDM_py_available()
}
\description{
is_sjSDM_py_available
check whetcher torch is available
}
