% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_torch_available}
\alias{is_torch_available}
\title{is_torch_available}
\usage{
is_torch_available()
}
\value{
Logical, is torch module available or not.
}
\description{
is_torch_available
}
\details{
check whether torch is available
}
