% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{missing_installation}
\alias{missing_installation}
\title{missing_installation}
\usage{
missing_installation(miss_torch, miss_sjSDM)
}
\arguments{
\item{miss_torch}{torch missing, logical}

\item{miss_sjSDM}{sjSDM_py missing, logical}
}
\description{
missing_installation
}
