% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\docType{data}
\name{torch}
\alias{torch}
\alias{fa}
\title{Torch module}
\format{
Torch module

fa module
}
\usage{
torch

fa
}
\description{
Interface to 'PyTorch' module

Interface to internal 'fa' module
}
\keyword{datasets}
