% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjSDM_configs.R
\name{multinomial}
\alias{multinomial}
\title{Multinomial}
\usage{
multinomial(invlink = "multinomial logit")
}
\arguments{
\item{invlink}{inverse link function, softmax is currently only supported}
}
\description{
multinomial family
}
