% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjSDM_cv.R
\name{new_image}
\alias{new_image}
\title{new_image function}
\usage{
new_image(
  z,
  cols = (grDevices::colorRampPalette(c("white", "#24526E"), bias = 1.5))(10),
  range = c(0.5, 1)
)
}
\arguments{
\item{z}{z matrix}

\item{cols}{cols for gradient}

\item{range}{rescale to range}
}
\description{
new_image function
}
