% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn.R
\name{optimizer_adamax}
\alias{optimizer_adamax}
\title{Adamax optimizer}
\usage{
optimizer_adamax(learning_rate = 0.01, weight_decay = 0.01)
}
\arguments{
\item{learning_rate}{learning_rate}

\item{weight_decay}{weight decay for optimizer}
}
\description{
create instance of adamax optimizer
}
