% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn.R
\name{optimizer_RMSprop}
\alias{optimizer_RMSprop}
\title{RMSprop optimizer}
\usage{
optimizer_RMSprop(learning_rate = 0.01, weight_decay = 0.01, ...)
}
\arguments{
\item{learning_rate}{learning_rate}

\item{weight_decay}{weight decay for optimizer}

\item{...}{additional arguments passed to RMSprop}
}
\description{
create instance of adamax optimizer
}
