% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning.R
\name{plot.sjSDM_cv}
\alias{plot.sjSDM_cv}
\title{Plot elastic net tuning}
\usage{
\method{plot}{sjSDM_cv}(x, y, perf = c("logLik", "AUC", "AUC_macro"), resolution = 6, k = 3, ...)
}
\arguments{
\item{x}{a model fitted by \code{\link{sjSDM_cv}}}

\item{y}{unused argument}

\item{perf}{performance measurement to plot}

\item{resolution}{resolution of grid}

\item{k}{number of knots for the gm}

\item{...}{Additional arguments to pass to \code{plot()}}
}
\description{
Plot elastic net tuning
}
