% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjSDM_DNN.R
\name{plot.sjSDM_DNN}
\alias{plot.sjSDM_DNN}
\title{Plot training history}
\usage{
\method{plot}{sjSDM_DNN}(x, y, ...)
}
\arguments{
\item{x}{a model fitted by \code{\link{sjSDM_DNN}}}

\item{y}{unused argument}

\item{...}{Additional arguments to pass to \code{plot()}}
}
\description{
Plot training history
}
