% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjSDM_model.R
\name{plot.sjSDM_model}
\alias{plot.sjSDM_model}
\title{Plot training history}
\usage{
\method{plot}{sjSDM_model}(x, y, ...)
}
\arguments{
\item{x}{a model fitted by \code{\link{sjSDM_model}}}

\item{y}{unused argument}

\item{...}{Additional arguments to pass to \code{plot()}}
}
\description{
Plot training history
}
