% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{plot.sjSDManovaIndividual}
\alias{plot.sjSDManovaIndividual}
\title{Plot anova}
\usage{
\method{plot}{sjSDManovaIndividual}(x, y, perf = c("LogLik", "R2"),
  cols = c("#7FC97F", "#BEAED4", "#FDC086"), alpha = 0.15, percent = TRUE,
  ...)
}
\arguments{
\item{x}{anova object from \code{\link{anova.sjSDM}}}

\item{y}{unused argument}

\item{perf}{performance measurement to plot}

\item{cols}{colors for the groups}

\item{alpha}{alpha for colors}

\item{percent}{use relative instead of absolute values}

\item{...}{Additional arguments to pass to \code{plot()}}
}
\description{
Plot anova
}
