% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalStructure.R
\name{plot.sjSDMinternalStruture}
\alias{plot.sjSDMinternalStruture}
\title{Plot internal structure}
\usage{
\method{plot}{sjSDMinternalStruture}(x, alpha = 0.15, env_deviance = NULL, ...)
}
\arguments{
\item{object}{and object of class sjSDMinternalStruture create by anova object from \code{\link{internalStructure}}}
}
\description{
Creates a ternary diagram of an object of class
}
