% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sLVM.R
\name{plot.sLVM}
\alias{plot.sLVM}
\title{sLVM ordinations plot}
\usage{
\method{plot}{sLVM}(x, ...)
}
\arguments{
\item{x}{an object of class \code{\link{sLVM}}}

\item{...}{optional arguments for compatibility with the generic function, passed to the plot function}
}
\description{
This S3 function creates the ordination plot for the LVM model
}
