% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalStructure.R
\name{plotAssemblyEffects}
\alias{plotAssemblyEffects}
\title{Plot correlations between assembly processes and predictors or traits}
\usage{
plotAssemblyEffects(
  object,
  env = NULL,
  trait = NULL,
  cols = c("#A38310", "#B42398", "#20A382"),
  negatives = c("floor", "scale", "raw")
)
}
\arguments{
\item{object}{An \code{sjSDManova} object from the \code{\link{anova.sjSDM}} function.}

\item{env}{Predictor variable. If \code{NULL}, assembly processes are plotted against environment, spatial uniqueness, and richness.}

\item{trait}{Trait variable. Plotted against species R-squared for the three processes.}

\item{cols}{Colors for the three assembly processes.}

\item{negatives}{how to handle negative R squareds

Correlation and plots of the three assembly processes (environment, space, and codist) against environmental and spatial uniqueness and richness. The importance of the three assembly processes is measured by the partial R-squared (shown in the internal structure plots).
Importances are available for species and sites. Custom environmental predictors or traits can be specified. Environmental predictors are plotted against site R-squared and traits are plotted against species R-squared.
Regression lines are estimated by 50\% quantile regression models.}
}
\value{
A list with the following components:

\item{env}{A list of summary tables for env, space, and codist R-squared.}
\item{space}{A list of summary tables for env, space, and codist R-squared.}
\item{codist}{A list of summary tables for env, space, and codist R-squared.}
}
\description{
Plot correlations between assembly processes and predictors or traits
}
\examples{
library(sjSDM)
# simulate community:
community = simulate_SDM(env = 3L, species = 10L, sites = 100L)

Occ <- community$response
Env <- community$env_weights
SP <- data.frame(matrix(rnorm(200, 0, 0.3), 100, 2)) # spatial coordinates


# fit model:
model <- sjSDM(Y = Occ, 
               env = linear(data = Env, formula = ~X1+X2+X3), 
               spatial = linear(data = SP, formula = ~0+X1*X2), 
               family=binomial("probit"),
               iter = 20) # increase iter for real analysis

# Calculate ANOVA for env, space, associations, for details see ?anova.sjSDM
an = anova(model, samples = 10) # increase iter for real analysis

# Show anova fractions
plot(an)

# ANOVA tables with different way to handle fractions
summary(an)
summary(an, fractions = "discard")
summary(an, fractions = "proportional")
summary(an, fractions = "equal")

# Internal structure
int = internalStructure(an, fractions = "equal")

print(int)

plot(int)

plotAssemblyEffects(int)
plotAssemblyEffects(int, env = runif(100))
plotAssemblyEffects(int, env = as.factor(c(rep(1, 50), rep(2, 50))))
plotAssemblyEffects(int, trait = runif(10))
plotAssemblyEffects(int, trait = as.factor(c(rep(1, 5), rep(2, 5))))

}
\references{
Leibold, M. A., Rudolph, F. J., Blanchet, F. G., De Meester, L., Gravel, D., Hartig, F., ... & Chase, J. M. (2022). The internal structure of metacommunities. *Oikos*, 2022(1).
}
