% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotAssociations}
\alias{plotAssociations}
\title{plotAssociations
plot species-species associations}
\usage{
plotAssociations(sigma, radius = 5, main = NULL, circleBreak = FALSE,
  top = 30L, occ = NULL)
}
\arguments{
\item{sigma}{species-species covariance matrix}

\item{radius}{circle's radius}

\item{main}{title}

\item{circleBreak}{circle break or not}

\item{top}{top negative and positive associations}

\item{occ}{species occurence data}
}
\description{
plotAssociations
plot species-species associations
}
