% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjSDM_linear.R
\name{predict.sjSDM.spatialRE}
\alias{predict.sjSDM.spatialRE}
\title{Predict from a fitted sjSDM model}
\usage{
\method{predict}{sjSDM.spatialRE}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{a model fitted by \code{\link{sjSDM}} with \code{\link{spatialRE}}}

\item{newdata}{newdata for predictions}

\item{...}{optional arguments for compatibility with the generic function, no function implemented}
}
\description{
Predict from a fitted sjSDM model
}
