% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjSDM_DNN.R
\name{predict.sjSDM_DNN}
\alias{predict.sjSDM_DNN}
\title{Predict from a fitted sjSDM DNN}
\usage{
\method{predict}{sjSDM_DNN}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{a model fitted by \code{\link{sjSDM_DNN}}}

\item{newdata}{newdata for predictions}

\item{...}{optional arguments for compatibility with the generic function, no function implemented}
}
\description{
Predict from a fitted sjSDM DNN
}
