% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjSDM_model.R
\name{predict.sjSDM_model}
\alias{predict.sjSDM_model}
\title{Predict from a fitted sjSDM_model}
\usage{
\method{predict}{sjSDM_model}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{a model fitted by \code{\link{sjSDM_model}}}

\item{newdata}{newdata for predictions}

\item{...}{optional arguments such as batch_size}
}
\description{
Predict from a fitted sjSDM_model
}
