% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sLVM.R
\name{predict.sLVM}
\alias{predict.sLVM}
\title{Predict from a fitted sLVM model}
\usage{
\method{predict}{sLVM}(object, newdata = NULL, mean_field = FALSE, ...)
}
\arguments{
\item{object}{a model fitted by \code{\link{sLVM}}}

\item{newdata}{newdata for predictions}

\item{mean_field}{use means of parameter estimates or samples from their distribution}

\item{...}{optional arguments for compatibility with the generic function, no function implemented}
}
\description{
Predict from a fitted sLVM model
}
