% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjSDM_configs.R
\name{print.bioticStruct}
\alias{print.bioticStruct}
\title{Print a bioticStruct object}
\usage{
\method{print}{bioticStruct}(x, ...)
}
\arguments{
\item{x}{object created by \code{\link{bioticStruct}}}

\item{...}{optional arguments for compatibility with the generic function, no function implemented}
}
\description{
Print a bioticStruct object
}
