% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjSDM_configs.R
\name{print.DNN}
\alias{print.DNN}
\title{Print a DNN object}
\usage{
\method{print}{DNN}(x, ...)
}
\arguments{
\item{x}{object created by \code{\link{DNN}}}

\item{...}{optional arguments for compatibility with the generic function, no function implemented}
}
\description{
Print a DNN object
}
