% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjSDM_configs.R
\name{print.envLinear}
\alias{print.envLinear}
\title{Print a envLinear object}
\usage{
\method{print}{envLinear}(x, ...)
}
\arguments{
\item{x}{object created by \code{\link{envLinear}}}

\item{...}{optional arguments for compatibility with the generic function, no function implemented}
}
\description{
Print a envLinear object
}
