% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjSDM_configs.R
\name{print.linear}
\alias{print.linear}
\title{Print a linear object}
\usage{
\method{print}{linear}(x, ...)
}
\arguments{
\item{x}{object created by \code{\link{linear}}}

\item{...}{optional arguments for compatibility with the generic function, no function implemented}
}
\value{
Invisible formula object
}
\description{
Print a linear object
}
