% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_blocks.R
\name{print.NN}
\alias{print.NN}
\title{Print an NN object}
\usage{
\method{print}{NN}(x, ...)
}
\arguments{
\item{x}{object created by \code{\link{NN}}}

\item{...}{optional arguments for compatibility with the generic function, no function implemented}
}
\value{
Invisible NN object
}
\description{
Print an NN object
}
