% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjSDM.R
\name{print.sjSDM}
\alias{print.sjSDM}
\title{Print a fitted sjSDM model}
\usage{
\method{print}{sjSDM}(x, ...)
}
\arguments{
\item{x}{a model fitted by \code{\link{sjSDM}}}

\item{...}{optional arguments for compatibility with the generic function, no function implemented}
}
\value{
No return value
}
\description{
Print a fitted sjSDM model
}
