% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjSDM_cv.R
\name{print.sjSDM_cv}
\alias{print.sjSDM_cv}
\title{Print a fitted sjSDM_cv model}
\usage{
\method{print}{sjSDM_cv}(x, ...)
}
\arguments{
\item{x}{a model fitted by \code{\link{sjSDM_cv}}}

\item{...}{optional arguments for compatibility with the generic function, no function implemented}
}
\value{
Above data frame is silently returned.
}
\description{
Print a fitted sjSDM_cv model
}
