% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjSDM_DNN.R
\name{print.sjSDM_DNN}
\alias{print.sjSDM_DNN}
\title{Print a fitted sjSDM DNN}
\usage{
\method{print}{sjSDM_DNN}(x, ...)
}
\arguments{
\item{x}{a moel fitted by \code{\link{sjSDM}} with \code{\link{envDNN}} object}

\item{...}{optional arguments for compatibility with the generic function, no function implemented}
}
\description{
Print a fitted sjSDM DNN
}
