% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjSDM_model.R
\name{print.sjSDM_model}
\alias{print.sjSDM_model}
\title{Print a fitted sjSDM model}
\usage{
\method{print}{sjSDM_model}(x, ...)
}
\arguments{
\item{x}{a moel fitted by \code{\link{sjSDM_model}}}

\item{...}{optional arguments for compatibility with the generic function, no function implemented}
}
\description{
Print a fitted sjSDM model
}
