% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{print.sjSDManova}
\alias{print.sjSDManova}
\title{Print sjSDM anova}
\usage{
\method{print}{sjSDManova}(x, ...)
}
\arguments{
\item{x}{an object of \code{\link{anova.sjSDM}}}

\item{...}{optional arguments for compatibility with the generic function, no function implemented}
}
\value{
The above matrix is silently returned
}
\description{
Print sjSDM anova
}
