% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalStructure.R
\name{print.sjSDMinternalStructure}
\alias{print.sjSDMinternalStructure}
\title{Print internal structure object}
\usage{
\method{print}{sjSDMinternalStructure}(x, ...)
}
\arguments{
\item{x}{object of class sjSDMinternalStructure}

\item{...}{no function}
}
\description{
Print internal structure object
}
