% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalStructure.R
\name{print.sjSDMinternalStruture}
\alias{print.sjSDMinternalStruture}
\title{Print internal structure object}
\usage{
\method{print}{sjSDMinternalStruture}(x, ...)
}
\arguments{
\item{x}{object of class sjSDMinternalStruture}

\item{...}{no function}
}
\description{
Print internal structure object
}
