% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sLVM.R
\name{print.sLVM}
\alias{print.sLVM}
\title{Print a fitted sLVM model}
\usage{
\method{print}{sLVM}(x, ...)
}
\arguments{
\item{x}{a model fitted by \code{\link{sLVM}}}

\item{...}{optional arguments for compatibility with the generic function, no function implemented}
}
\description{
Print a fitted sLVM model
}
