% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjSDM_configs.R
\name{print.spatialRE}
\alias{print.spatialRE}
\title{Print a spatialRE object}
\usage{
\method{print}{spatialRE}(x, ...)
}
\arguments{
\item{x}{object created by \code{\link{spatialRE}}}

\item{...}{optional arguments for compatibility with the generic function, no function implemented}
}
\description{
Print a spatialRE object
}
