% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.R
\name{py_install_method_detect}
\alias{py_install_method_detect}
\title{py_install_method_detect}
\usage{
py_install_method_detect(envname, conda = "auto")
}
\arguments{
\item{envname}{env name}

\item{conda}{conda path}
}
\description{
py_install_method_detect from `reticulate::py_install_method_detect()`
}
