% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjSDM_linear.R
\name{ranef}
\alias{ranef}
\title{Extract random effects}
\usage{
ranef(object)
}
\arguments{
\item{object}{a model fitted by \code{\link{sjSDM}} with \code{\link{spatialRE}}}
}
\description{
Extract random effects
}
