% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjSDM.R
\name{residuals.sjSDM}
\alias{residuals.sjSDM}
\title{Residuals for a sjSDM model}
\usage{
\method{residuals}{sjSDM}(object, type = "raw", ...)
}
\arguments{
\item{object}{a model fitted by \code{\link{sjSDM}}}

\item{type}{residual type. Currently only supports raw}

\item{...}{further arguments, not supported yet.}
}
\value{
residuals in the format of the provided community matrix
}
\description{
Returns residuals for a fitted sjSDM model
}
