% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjSDM_configs.R
\name{RMSprop}
\alias{RMSprop}
\title{RMSprop}
\usage{
RMSprop(
  alpha = 0.99,
  eps = 1e-08,
  weight_decay = 0.01,
  momentum = 0.1,
  centered = FALSE
)
}
\arguments{
\item{alpha}{decay factor}

\item{eps}{fuzz factor}

\item{weight_decay}{l2 penalty on weights}

\item{momentum}{momentum}

\item{centered}{centered or not}
}
\value{
Anonymous function that returns optimizer when called.
}
\description{
RMSprop optimizer
}
