% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsquared.R
\name{Rsquared}
\alias{Rsquared}
\title{Rsquared}
\usage{
Rsquared(model, X = NULL, Y = NULL, SP = NULL, adjust = FALSE,
  averageSP = TRUE, averageSite = TRUE)
}
\arguments{
\item{model}{model}

\item{X}{new environmental covariates}

\item{Y}{new species occurrences}

\item{SP}{new spatial covariates}

\item{adjust}{adjust R squared or not}

\item{averageSP}{average R squared over species}

\item{averageSite}{average R squared over sites}
}
\description{
calculate Rsquared following Nakagawa
}
\author{
Maximilian Pichler
}
