% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsquared.R
\name{Rsquared2}
\alias{Rsquared2}
\title{Rsquared2}
\usage{
Rsquared2(model, X = NULL, Y = NULL, SP = NULL, individual = TRUE, ...)
}
\arguments{
\item{model}{model}

\item{X}{new environmental covariates}

\item{Y}{new species occurences}

\item{SP}{new spatial covariates}

\item{individual}{R squared for each site}

\item{...}{additional parameters}
}
\description{
claculate Rsquared following Nakagawa
}
\author{
Maximilian Pichler
}
