% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjSDM_model.R
\name{setWeights}
\alias{setWeights}
\title{Set weights}
\usage{
setWeights(object, weights)
}
\arguments{
\item{object}{object of class \code{\link{sjSDM_model}}}

\item{weights}{list of layer weights and sigma, see \code{\link{getWeights}}}
}
\description{
set layer weights and sigma in \code{\link{sjSDM_model}} object
}
