% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjSDM_configs.R
\name{SGD}
\alias{SGD}
\title{SGD}
\usage{
SGD(momentum = 0.5, dampening = 0, weight_decay = 0, nesterov = TRUE)
}
\arguments{
\item{momentum}{betas}

\item{dampening}{eps}

\item{weight_decay}{weight_decay}

\item{nesterov}{nesterov}
}
\description{
SGD
}
