% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjSDM.R
\name{simulate.sjSDM}
\alias{simulate.sjSDM}
\title{Generates simulations from sjSDM model}
\usage{
\method{simulate}{sjSDM}(object, nsim = 1, seed = NULL, ...)
}
\arguments{
\item{object}{object of class sjSDM}

\item{nsim}{number of simulations}

\item{seed}{seed for random numer generator}

\item{...}{additional parameters}
}
\description{
Simulate nsim responses from the fitted model
}
