% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjSDM_linear.R
\name{simulate.sjSDM.spatialRE}
\alias{simulate.sjSDM.spatialRE}
\title{Generates simulations from sjSDM model}
\usage{
\method{simulate}{sjSDM.spatialRE}(object, nsim = 1, seed = NULL, ...)
}
\arguments{
\item{object}{a model fitted by \code{\link{sjSDM}} with \code{\link{spatialRE}}}

\item{nsim}{number of simulations}

\item{seed}{seed for random numer generator}

\item{...}{optional arguments for compatibility with the generic function, no functionality implemented}
}
\description{
Simulate nsim responses from the fitted model
}
