% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_SDM}
\alias{simulate_SDM}
\title{Simulate joint Species Distribution Models}
\usage{
simulate_SDM(env = 5L, sites = 100L, species = 5L, Re = NULL,
  correlation = TRUE, weight_range = c(-1, 1), link = "probit",
  response = "pa", sparse = NULL, tolerance = 0.05, iter = 20L,
  seed = NULL)
}
\arguments{
\item{env}{number of environment variables}

\item{sites}{number of sites}

\item{species}{number of species}

\item{Re}{random effects (intercepts)}

\item{correlation}{correlated species TRUE or FALSE, can be also a function or a matrix}

\item{weight_range}{sample true weights from uniform range, default -1,1}

\item{link}{probit, logit or idential}

\item{response}{pa (presence-absence) or count}

\item{sparse}{sparse rate}

\item{tolerance}{tolerance for sparsity check}

\item{iter}{tries until sparse rate is achieved}

\item{seed}{random seed. Default = 42}
}
\description{
Simulate species distributions
}
\details{
probit is not possible for response = count
}
\author{
Maximilian Pichler
}
