% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn.R
\name{sjSDM_model}
\alias{sjSDM_model}
\title{sjSDM_model}
\usage{
sjSDM_model(input_shape, device = NULL, dtype = "float32")
}
\arguments{
\item{input_shape}{number of predictors}

\item{device}{which device to be used, "cpu" or "gpu"}

\item{dtype}{which data type, most GPU support only 32 bit floats.}
}
\description{
Create a new sjSDM_model
}
