% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sLVM.R
\name{sLVM}
\alias{sLVM}
\title{sLVM
scalable LVM model}
\usage{
sLVM(Y = NULL, X = NULL, formula = NULL, lv = 2L, family,
  priors = list(3, 1, 1), posterior = c("DiagonalNormal",
  "LaplaceApproximation", "LowRankMultivariateNormal", "Delta"), iter = 50L,
  step_size = 20L, lr = list(0.1), device = "cpu", dtype = "float32")
}
\arguments{
\item{Y}{species occurrences}

\item{X}{environmental (abiotic) covariates}

\item{formula}{formula for environment}

\item{lv}{number of latent variables}

\item{family}{supported distributions: \code{binomial(link=c("logit", "probit"))}, \code{poisson(link=c("log", "identity"))}}

\item{priors}{list of scale priors for beta, lv, and lf}

\item{posterior}{type of posterior distribution}

\item{iter}{number of optimization steps}

\item{step_size}{batch_size}

\item{lr}{learning_rate, can be also a list (for each parameter type)}
}
\description{
sLVM
scalable LVM model
}
