% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjSDM.R
\name{summary.sjSDM}
\alias{summary.sjSDM}
\title{Return summary of a fitted sjSDM model}
\usage{
\method{summary}{sjSDM}(object, ...)
}
\arguments{
\item{...}{optional arguments for compatibility with the generic function, no function implemented}

\item{x}{a model fitted by \code{\link{sjSDM}}}
}
\description{
Return summary of a fitted sjSDM model
}
