% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjSDM_DNN.R
\name{summary.sjSDM_DNN}
\alias{summary.sjSDM_DNN}
\title{Return summary of a fitted sjSDM DNN}
\usage{
\method{summary}{sjSDM_DNN}(object, ...)
}
\arguments{
\item{object}{a model fitted by \code{\link{sjSDM}} with \code{\link{envDNN}} object}

\item{...}{optional arguments for compatibility with the generic function, no functionality implemented}
}
\description{
Return summary of a fitted sjSDM DNN
}
