% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjSDM_model.R
\name{summary.sjSDM_model}
\alias{summary.sjSDM_model}
\title{Return summary of a fitted sjSDM model}
\usage{
\method{summary}{sjSDM_model}(object, ...)
}
\arguments{
\item{object}{a model fitted by \code{\link{sjSDM_model}}}

\item{...}{optional arguments for compatibility with the generic function, no functionality implemented}
}
\description{
Return summary of a fitted sjSDM model
}
