% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sLVM.R
\name{summary.sLVM}
\alias{summary.sLVM}
\title{Return summary of a fitted sLVM model}
\usage{
\method{summary}{sLVM}(object, ...)
}
\arguments{
\item{object}{a model fitted by \code{\link{sLVM}}}

\item{...}{optional arguments for compatibility with the generic function, no functionality implemented}
}
\description{
Return summary of a fitted sLVM model
}
