% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varPart.R
\name{varPart}
\alias{varPart}
\title{varPart}
\usage{
varPart(beta, sp = NULL, covariance, covX, covSP = NULL)
}
\arguments{
\item{beta}{abiotic weights}

\item{sp}{spatial weights}

\item{covariance}{species associations}

\item{covX}{environmental covariance matrix}

\item{covSP}{spatial covariance matrix}
}
\description{
variation partitioning
}
\author{
Maximilian Pichler
}
