% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{varPartTypIII}
\alias{varPartTypIII}
\title{varPartTypIII}
\usage{
varPartTypIII(model, order = NULL, ...)
}
\arguments{
\item{model}{object fitted by \code{\link{sjSDM}}}

\item{order}{which modules should be removed: E (environment), S (spatial), or B (biotic)}

\item{...}{arguments passed to \code{\link{Rsquared}}}
}
\description{
variation partitioning Typ III
}
\author{
Maximilian Pichler
}
