% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.R
\name{virtualenv_default_python}
\alias{virtualenv_default_python}
\title{virtualenv_default_python from reticulate}
\usage{
virtualenv_default_python(python = NULL)
}
\description{
virtualenv_default_python from reticulate
}
