% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjSDM_configs.R
\name{AccSGD}
\alias{AccSGD}
\title{AccSGD}
\usage{
AccSGD(kappa = 1000, xi = 10, small_const = 0.7, weight_decay = 0)
}
\arguments{
\item{kappa}{betas}

\item{xi}{eps}

\item{small_const}{small_const}

\item{weight_decay}{weight_decay}
}
\description{
AccSGD
}
