% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova.sjSDM}
\alias{anova.sjSDM}
\title{Anova}
\usage{
\method{anova}{sjSDM}(object, samples = 5000L, ...)
}
\arguments{
\item{object}{model of object \code{\link{sjSDM}}}

\item{samples}{Number of Monte Carlo samples}

\item{...}{optional arguments which are passed to the calculation of the logLikelihood}

\item{fractions}{how to distribute the shared partitions, proportional to the fractions' R2s or equal}
}
\value{
An S3 class of type 'sjSDManova' including the following components:

\item{results}{Data frame of results.}
\item{to_print}{Data frame, summarized results for type I anova.}
\item{N}{Number of observations (sites).}
\item{spatial}{Logical, spatial model or not.}
\item{species}{individual species R2s.}
\item{sites}{individual site R2s.}
\item{lls}{individual site by species negative-log-likelihood values.}
\item{model}{model}

Implemented S3 methods are \code{\link{print.sjSDManova}} and \code{\link{plot.sjSDManova}}
}
\description{
Calculates type II anova.

Shared contributions (e.g. between space and environment) are also calculated (and divided proportionally) and can be optionally visualized via \code{\link{plot.sjSDManova}} with \code{add_shared=TRUE}.
The anova can get unstable for many species and few occurrences/observations. We recommend using large numbers for 'samples'.
}
\details{
Compute analysis of variance
}
\examples{
\dontrun{

library(sjSDM)
# simulate community:
com = simulate_SDM(env = 3L, species = 10L, sites = 100L)
XY = matrix(runif(100, -1, 1), 100, 2)

# fit model:
model = sjSDM(Y = com$response,env = com$env_weights,spatial = linear(XY, ~0+.), iter = 50L) 
# increase iter for your own data 

# Anova 
an = anova(model)
plot(an)

print(an, fractions = "equal")
print(an, fractions = "proportional")
print(an, fractions = "discard")




# Internal structure
plotInternalStructure(an, fractions = "discard")

plotAssemblyEffects(an)
plotAssemblyEffects(an, env = runif(100))
plotAssemblyEffects(an, env = as.factor(c(rep(1, 50), rep(2, 50))))
plotAssemblyEffects(an, trait = runif(10))
plotAssemblyEffects(an, trait = as.factor(c(rep(1, 5), rep(2, 5))))


}
}
\seealso{
\code{\link{plot.sjSDManova}}, \code{\link{print.sjSDManova}}, \code{\link{plotInternalStructure}}
}
