% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova.sjSDM}
\alias{anova.sjSDM}
\title{Anova}
\usage{
\method{anova}{sjSDM}(object, cv = 5L, ...)
}
\arguments{
\item{object}{model of object \code{\link{sjSDM}}}

\item{cv}{number of cross-validation splits}

\item{...}{optional arguments for compatibility with the generic function, no function implemented}
}
\description{
Compute analysis of variance via cross-validation for environmental, associations, and spatial effects
}
\seealso{
\code{\link{plot.sjSDManova}}, \code{\link{print.sjSDManova}}
}
