% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjSDM_configs.R
\name{DNN}
\alias{DNN}
\title{env DNN}
\usage{
DNN(data = NULL, formula = NULL, hidden = c(10L, 10L, 10L),
  activation = "relu", lambda = 0, alpha = 0.5)
}
\arguments{
\item{data}{matrix of environmental predictors}

\item{formula}{formula object for predictors}

\item{hidden}{hidden units in layers, length of hidden correspond to number of layers}

\item{activation}{activation functions for layer, must be of same length as hidden}

\item{lambda}{lambda penality}

\item{alpha}{weighting between LASSO and ridge}
}
\description{
create deep neural network
}
