% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generateSpatialEV}
\alias{generateSpatialEV}
\title{Generate spatial eigenvectors}
\usage{
generateSpatialEV(coords = NULL, threshold = 0)
}
\arguments{
\item{coords}{matrix or data.frame of coordinates}

\item{threshold}{ignore distances greater than threshold}
}
\value{
Matrix of spatial eigenvectors.
}
\description{
function to generate spatial eigenvectors to account for spatial autocorrelation
}
