% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getCor}
\alias{getCor}
\alias{getCor.sjSDM}
\title{getCor}
\usage{
getCor(object)

\method{getCor}{sjSDM}(object)
}
\arguments{
\item{object}{a model fitted by \code{\link{sjSDM}}, or \code{\link{sjSDM}} with \code{\link{DNN}} object}
}
\value{
Matrix of dimensions species by species corresponding to the covariance (occurrence) matrix.
}
\description{
get species-species association correlation matrix
}
\seealso{
\code{\link{sjSDM}},\code{\link{DNN}}
}
