% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getCov}
\alias{getCov}
\alias{getCov.sjSDM}
\alias{getCov.sjSDM_model}
\title{getCov}
\usage{
getCov(object)

\method{getCov}{sjSDM}(object)

\method{getCov}{sjSDM_model}(object)
}
\arguments{
\item{object}{a model fitted by \code{\link{sjSDM}} or of class \code{\link{sjSDM_model}}}
}
\description{
get species-species assocation (covariance) matrix
}
\seealso{
\code{\link{sjSDM}}, \code{\link{sjSDM_model}}
}
