% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjSDM_model.R
\name{getWeights}
\alias{getWeights}
\title{Get weights}
\usage{
getWeights(object)
}
\arguments{
\item{object}{object of class \code{\link{sjSDM_model}}}
}
\value{
\itemize{
 \item layers - list of layer weights
 \item sigma - weight to construct covariance matrix
}
}
\description{
return weights of each layer
}
