% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getWeights}
\alias{getWeights}
\alias{getWeights.sjSDM}
\title{Get weights}
\usage{
getWeights(object)

\method{getWeights}{sjSDM}(object)
}
\arguments{
\item{object}{object of class \code{\link{sjSDM}} with \code{\link{DNN}}}
}
\value{
\itemize{
\item layers - list of layer weights
\item sigma - weight to construct covariance matrix
}
}
\description{
return weights of each layer
}
