% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_sjSDM}
\alias{install_sjSDM}
\title{Install sjSDM and its dependencies}
\usage{
install_sjSDM(method = "conda", conda = "auto", version = c("cpu", "gpu"),
  envname = "r-sjSDM", extra_packages = NULL, restart_session = TRUE,
  conda_python_version = "3.6", pip = FALSE, cuda = c("10.1", "9,2"), ...)
}
\arguments{
\item{method}{installation method, auto = automatically best (conda or virtualenv), or force conda with method = "conda" or virtualenv with method = "virtualenv"}

\item{conda}{path to conda}

\item{version}{version = "cpu" for CPU version, or "gpu" for gpu version. (note MacOS users have to install cuda binaries by themselves)}

\item{envname}{Name of python env, "r-pytorch" is default}

\item{extra_packages}{Additional Python packages to install along with
PyTorch}

\item{restart_session}{Restart R session after installing (note this will
only occur within RStudio).}

\item{conda_python_version}{python version to be installed in the env, default = 3.6}

\item{cuda}{which cuda version, 9.2 and 10.1 are supported}
}
\description{
Install sjSDM and its dependencies
}
