% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjSDM.R
\name{logLik.sjSDM}
\alias{logLik.sjSDM}
\title{Extract Log-Likelihood from a fitted sjSDM model}
\usage{
\method{logLik}{sjSDM}(object, ...)
}
\arguments{
\item{object}{a model fitted by \code{\link{sjSDM}}}

\item{...}{optional arguments for compatibility with the generic function, no functionality implemented}
}
\value{
Numeric value
}
\description{
Extract Log-Likelihood from a fitted sjSDM model
}
