% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjSDM.R
\name{logLik.sjSDM}
\alias{logLik.sjSDM}
\title{Extract negative-log-Likelihood from a fitted sjSDM model}
\usage{
\method{logLik}{sjSDM}(object, individual = FALSE, ...)
}
\arguments{
\item{object}{a model fitted by \code{\link{sjSDM}}}

\item{individual}{returns internal ll structure, mostly for internal useage}

\item{...}{optional arguments passed to internal logLik function (only used if \code{individual=TRUE})}
}
\value{
Numeric value or numeric matrix if individual is true.
}
\description{
Extract negative-log-Likelihood from a fitted sjSDM model
}
