% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjSDM_cv.R
\name{plot.sjSDM_cv}
\alias{plot.sjSDM_cv}
\title{Plot elastic net tuning}
\usage{
\method{plot}{sjSDM_cv}(x, y, perf = c("logLik", "AUC", "AUC_macro"), resolution = 6, k = 3, ...)
}
\arguments{
\item{x}{a model fitted by \code{\link{sjSDM_cv}}}

\item{y}{unused argument}

\item{perf}{performance measurement to plot}

\item{resolution}{resolution of grid}

\item{k}{number of knots for the gm}

\item{...}{Additional arguments to pass to \code{plot()}}
}
\value{
Named vector of optimized regularization parameters.

Without space:

\item{lambda_cov}{Regularization strength in the \code{\link{bioticStruct}} object.}
\item{alpha_cov}{Weigthing between L1 and L2 in the \code{\link{bioticStruct}} object.}
\item{lambda_coef}{Regularization strength in the \code{\link{linear}} or \code{\link{DNN}} object.}
\item{alpha_coef}{Weigthing between L1 and L2 in the \code{\link{linear}} or \code{\link{DNN}} object.}

With space:

\item{lambda_cov}{Regularization strength in the \code{\link{bioticStruct}} object.}
\item{alpha_cov}{Weigthing between L1 and L2 in the \code{\link{bioticStruct}} object.}
\item{lambda_coef}{Regularization strength in the \code{\link{linear}} or \code{\link{DNN}} object.}
\item{alpha_coef}{Weigthing between L1 and L2 in the \code{\link{linear}} or \code{\link{DNN}} object.}
\item{lambda_spatial}{Regularization strength in the \code{\link{linear}} or \code{\link{DNN}} object for the spatial component.}
\item{alpha_spatial}{Weigthing between L1 and L2 in the\code{\link{linear}} or \code{\link{DNN}} object for the spatial component.}
}
\description{
Plot elastic net tuning
}
