% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{plot.sjSDManova}
\alias{plot.sjSDManova}
\title{Plot anova results}
\usage{
\method{plot}{sjSDManova}(
  x,
  y,
  type = c("McFadden", "Deviance", "Nagelkerke"),
  fractions = c("discard", "proportional", "equal"),
  cols = c("#7FC97F", "#BEAED4", "#FDC086"),
  alpha = 0.15,
  env_deviance = NULL,
  ...
)
}
\arguments{
\item{x}{anova object from \code{\link{anova.sjSDM}}}

\item{y}{unused argument}

\item{type}{deviance, Nagelkerke or McFadden R-squared}

\item{fractions}{how to handle shared fractions}

\item{cols}{colors for the groups}

\item{alpha}{alpha for colors}

\item{env_deviance}{environmental deviance}

\item{...}{Additional arguments to pass to \code{plot()}}
}
\value{
List with the following components:

\item{VENN}{Matrix of shown results.}
}
\description{
Plot anova results
}
\references{
Leibold, M. A., Rudolph, F. J., Blanchet, F. G., De Meester, L., Gravel, D., Hartig, F., ... & Chase, J. M. (2022). The internal structure of metacommunities. Oikos, 2022(1).
}
