% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{plot.sjSDManova}
\alias{plot.sjSDManova}
\title{Plot anova results}
\usage{
\method{plot}{sjSDManova}(
  x,
  y,
  type = c("Deviance", "Nagelkerke", "McFadden"),
  cols = c("#7FC97F", "#BEAED4", "#FDC086"),
  alpha = 0.15,
  ...
)
}
\arguments{
\item{x}{anova object from \code{\link{anova.sjSDM}}}

\item{y}{unused argument}

\item{type}{use of deviance or of Nagelkerke or McFadden R-squared}

\item{cols}{colors for the groups}

\item{alpha}{alpha for colors}

\item{...}{Additional arguments to pass to \code{plot()}}
}
\value{
The visualized matrix is silently returned
}
\description{
Plot anova results
}
