% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{plot.sjSDManovaIndividual}
\alias{plot.sjSDManovaIndividual}
\title{Plot anova}
\usage{
\method{plot}{sjSDManovaIndividual}(
  x,
  y,
  contour = FALSE,
  col.points = "#24526e",
  cex.points = 1.2,
  pch = 19,
  ...
)
}
\arguments{
\item{x}{anova object from \code{\link{anova.sjSDM}}}

\item{y}{unused argument}

\item{contour}{plot contour or not}

\item{col.points}{point color}

\item{cex.points}{point size}

\item{pch}{point symbol}

\item{...}{Additional arguments to pass to \code{plot()}}
}
\description{
Plot anova
}
