% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{plot.sjSDMimportance}
\alias{plot.sjSDMimportance}
\title{Plot importance}
\usage{
\method{plot}{sjSDMimportance}(x, y, col.points = "#24526e", cex.points = 1.2, ...)
}
\arguments{
\item{x}{a model fitted by \code{\link{importance}}}

\item{y}{unused argument}

\item{col.points}{point color}

\item{cex.points}{point size}

\item{...}{Additional arguments to pass to \code{plot()}}
}
\value{
The visualized matrix is silently returned.
}
\description{
Plot importance
}
