% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalStructure.R
\name{plot.sjSDMinternalStruture}
\alias{plot.sjSDMinternalStruture}
\title{Plot internal structure}
\usage{
\method{plot}{sjSDMinternalStruture}(
  x,
  alpha = 0.15,
  env_deviance = NULL,
  negatives = c("floor", "scale", "raw"),
  ...
)
}
\arguments{
\item{x}{and object of class sjSDMinternalStruture create by anova object from \code{\link{internalStructure}}}

\item{alpha}{alpha of points}

\item{env_deviance}{environmental deviance/gradient (points will be colored)}

\item{negatives}{how to handle negative R squareds}

\item{...}{no function}
}
\description{
Creates a ternary diagram of an object of class
}
\examples{
library(sjSDM)
# simulate community:
community = simulate_SDM(env = 3L, species = 10L, sites = 100L)

Occ <- community$response
Env <- community$env_weights
SP <- data.frame(matrix(rnorm(200, 0, 0.3), 100, 2)) # spatial coordinates


# fit model:
model <- sjSDM(Y = Occ, 
               env = linear(data = Env, formula = ~X1+X2+X3), 
               spatial = linear(data = SP, formula = ~0+X1*X2), 
               family=binomial("probit"),
               iter = 20) # increase iter for real analysis

# Calculate ANOVA for env, space, associations, for details see ?anova.sjSDM
an = anova(model, samples = 10) # increase iter for real analysis

# Show anova fractions
plot(an)

# ANOVA tables with different way to handle fractions
summary(an)
summary(an, fractions = "discard")
summary(an, fractions = "proportional")
summary(an, fractions = "equal")

# Internal structure
int = internalStructure(an, fractions = "equal")

print(int)

plot(int)

plotAssemblyEffects(int)
plotAssemblyEffects(int, env = runif(100))
plotAssemblyEffects(int, env = as.factor(c(rep(1, 50), rep(2, 50))))
plotAssemblyEffects(int, trait = runif(10))
plotAssemblyEffects(int, trait = as.factor(c(rep(1, 5), rep(2, 5))))

}
