% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{plotAssemblyEffects}
\alias{plotAssemblyEffects}
\title{Plot Correlations between assembly processes and predictors or traits}
\usage{
plotAssemblyEffects(
  object,
  env = NULL,
  trait = NULL,
  Rsquared = c("McFadden", "Nagelkerke"),
  cols = c("#A38310", "#B42398", "#20A382")
)
}
\arguments{
\item{object}{An `anova` object from the `anova.sjSDM` function.}

\item{env}{Predictor variable. If `NULL`, assembly processes are plotted against environment, spatial uniqueness, and richness.}

\item{trait}{Trait variable. Plotted against species R-squared for the three processes.}

\item{Rsquared}{Which R-squared should be used: "McFadden" (default) or "Nagelkerke".}

\item{cols}{Colors for the three assembly processes.

Correlation and plots of the three assembly processes (environment, space, and codist) against environmental and spatial uniqueness and richness. The importance of the three assembly processes is measured by the partial R-squared (shown in the internal structure plots).

Importances are available for species and sites. Custom environmental predictors or traits can be specified. Environmental predictors are plotted against site R-squared and traits are plotted against species R-squared.

Regression lines are estimated by 50% quantile regression models.}
}
\value{
A list with the following components:

\item{env}{A list of summary tables for env, space, and codist R-squared.}
\item{space}{A list of summary tables for env, space, and codist R-squared.}
\item{codist}{A list of summary tables for env, space, and codist R-squared.}
}
\description{
Plot Correlations between assembly processes and predictors or traits
}
\references{
Leibold, M. A., Rudolph, F. J., Blanchet, F. G., De Meester, L., Gravel, D., Hartig, F., ... & Chase, J. M. (2022). The internal structure of metacommunities. *Oikos*, 2022(1).
}
