% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotAssociations}
\alias{plotAssociations}
\title{plotAssociations
plot species-species associations}
\usage{
plotAssociations(sigma, radius = 5, main = NULL, circleBreak = FALSE,
  top = 10L, occ = NULL, cols_association = c("#FF0000", "#BF003F",
  "#7F007F", "#3F00BF", "#0000FF"), cols_occurrence = c("#BEBEBE", "#8E8E8E",
  "#5F5F5F", "#2F2F2F", "#000000"), lwd_occurrence = 1,
  species_indices = NULL)
}
\arguments{
\item{sigma}{species-species covariance matrix}

\item{radius}{circle's radius}

\item{main}{title}

\item{circleBreak}{circle break or not}

\item{top}{top negative and positive associations}

\item{occ}{species occurence data}

\item{cols_association}{col gradient for association lines}

\item{cols_occurrence}{col gradient for species}

\item{lwd_occurrence}{lwd for occurrence lines}

\item{species_indices}{indices for sorting species}
}
\description{
plotAssociations
plot species-species associations
}
