% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotAssociations}
\alias{plotAssociations}
\title{plotAssociations
plot species-species associations}
\usage{
plotAssociations(model, radius = 5, main = NULL, circleBreak = FALSE)
}
\arguments{
\item{model}{model of `class(model) == 'fajsm'`, see \code{\link{fajsm}}}

\item{radius}{circle's radius}

\item{main}{title}

\item{circleBreak}{circle break or not}
}
\description{
plotAssociations
plot species-species associations
}
