% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjSDM.R
\name{predict.sjSDM}
\alias{predict.sjSDM}
\title{Predict from a fitted sjSDM model}
\usage{
\method{predict}{sjSDM}(
  object,
  newdata = NULL,
  SP = NULL,
  Y = NULL,
  type = c("link", "raw"),
  dropout = FALSE,
  ...
)
}
\arguments{
\item{object}{a model fitted by \code{\link{sjSDM}}}

\item{newdata}{newdata for predictions}

\item{SP}{spatial predictors (e.g. X and Y coordinates)}

\item{Y}{Known occurrences of species, must be a matrix of the original size, species to be predicted must consist of NAs}

\item{type}{raw or link}

\item{dropout}{use dropout for predictions or not, only supported for DNNs}

\item{...}{optional arguments for compatibility with the generic function, no function implemented}
}
\value{
Matrix of predictions (sites by species)
}
\description{
Predict from a fitted sjSDM model
}
