% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjSDM.R
\name{predict.sjSDM}
\alias{predict.sjSDM}
\title{Predict from a fitted sjSDM model}
\usage{
\method{predict}{sjSDM}(object, newdata = NULL, ...)
}
\arguments{
\item{...}{optional arguments for compatibility with the generic function, no function implemented}

\item{x}{a model fitted by \code{\link{sjSDM}}}
}
\description{
Predict from a fitted sjSDM model
}
