% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjSDM.R
\name{predict.sjSDM}
\alias{predict.sjSDM}
\title{Predict from a fitted sjSDM model}
\usage{
\method{predict}{sjSDM}(object, newdata = NULL, SP = NULL, ...)
}
\arguments{
\item{object}{a model fitted by \code{\link{sjSDM}}}

\item{newdata}{newdata for predictions}

\item{SP}{spatial predictors (e.g. X and Y coordinates)}

\item{...}{optional arguments for compatibility with the generic function, no function implemented}
}
\description{
Predict from a fitted sjSDM model
}
