% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{print.sjSDMimportance}
\alias{print.sjSDMimportance}
\title{Print importance}
\usage{
\method{print}{sjSDMimportance}(x, ...)
}
\arguments{
\item{x}{an object of \code{\link{importance}}}

\item{...}{optional arguments for compatibility with the generic function, no function implemented}
}
\value{
The matrix above is silently returned
}
\description{
Print importance
}
