% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsquared.R
\name{Rsquared}
\alias{Rsquared}
\title{R-squared}
\usage{
Rsquared(model, method = c("McFadden", "Nagelkerke"), verbose = TRUE)
}
\arguments{
\item{model}{model}

\item{method}{McFadden or Nagelkerke}

\item{verbose}{\code{TRUE} or \code{FALSE}, indicating whether progress should be printed or not}
}
\value{
R-squared as numeric value
}
\description{
calculate R-squared following McFadden or Nagelkerke
}
\details{
\loadmathjax

Calculate R-squared following Nagelkerke or McFadden:

\itemize{
\item Nagelkerke: \mjseqn{R^2 = 1 - \exp(2/N \cdot (log\mathcal{L}_0 - log\mathcal{L}_1 ) )}
\item McFadden: \mjseqn{R^2 = 1 - log\mathcal{L}_1 / log\mathcal{L}_0  }
}
}
\author{
Maximilian Pichler
}
