% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsquared.R
\name{Rsquared}
\alias{Rsquared}
\title{R-squared}
\usage{
Rsquared(model, method = c("Nagelkerke", "McFadden"))
}
\arguments{
\item{model}{model}

\item{method}{Nagelkerke or McFadden}
}
\value{
R-squared as numeric value
}
\description{
calculate R-squared following Nagelkerke or McFadden
}
\details{
\loadmathjax 

Calculate R-squared following Nagelkerke or McFadden:

\itemize{
\item Nagelkerke: \mjeqn{R^2 = 1 - \exp(2/N \cdot (log\mathcal{L}_0 - log\mathcal{L}_1 ) )}{}
\item McFadden: \mjeqn{R^2 = 1 - log\mathcal{L}_1 / log\mathcal{L}_0  }{}
}
}
\author{
Maximilian Pichler
}
