% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{setWeights}
\alias{setWeights}
\alias{setWeights.sjSDM_model}
\alias{setWeights.sjSDM_DNN}
\title{Set weights}
\usage{
setWeights(object, weights)

\method{setWeights}{sjSDM_model}(object, weights)

\method{setWeights}{sjSDM_DNN}(object, weights = NULL)
}
\arguments{
\item{object}{object of class \code{\link{sjSDM_model}} or \code{\link{sjSDM}} with \code{\link{envDNN}} object}

\item{weights}{list of layer weights and sigma, see \code{\link{getWeights}}}
}
\description{
set layer weights and sigma in \code{\link{sjSDM_model}} or \code{\link{sjSDM}} with \code{\link{envDNN}} object
}
