% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{setWeights}
\alias{setWeights}
\alias{setWeights.sjSDM}
\title{Set weights}
\usage{
setWeights(object, weights)

\method{setWeights}{sjSDM}(object, weights = NULL)
}
\arguments{
\item{object}{object of class  \code{\link{sjSDM}} with \code{\link{DNN}} object}

\item{weights}{list of layer weights and sigma, see \code{\link{getWeights}}}
}
\value{
No return value, weights are changed in place.
}
\description{
set layer weights and sigma in \code{\link{sjSDM}} with \code{\link{DNN}} object
}
