% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjSDM_configs.R
\name{SGD}
\alias{SGD}
\title{SGD}
\usage{
SGD(momentum = 0.5, dampening = 0, weight_decay = 0, nesterov = TRUE)
}
\arguments{
\item{momentum}{strength of momentum}

\item{dampening}{decay}

\item{weight_decay}{l2 penalty on weights}

\item{nesterov}{Nesterov momentum or not}
}
\value{
Anonymous function that returns optimizer when called.
}
\description{
stochastic gradient descent optimizer
}
