% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjSDM.R
\name{simulate.sjSDM}
\alias{simulate.sjSDM}
\title{Generates simulations from sjSDM model}
\usage{
\method{simulate}{sjSDM}(object, nsim = 1, seed = NULL, ...)
}
\arguments{
\item{object}{a model fitted by \code{\link{sjSDM}}}

\item{nsim}{number of simulations}

\item{seed}{seed for random numer generator}

\item{...}{optional arguments for compatibility with the generic function, no functionality implemented}
}
\value{
Array of simulated species occurrences.
}
\description{
Simulate nsim responses from the fitted model
}
