% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjSDM.R
\name{sjSDM}
\alias{sjSDM}
\title{sjSDM}
\usage{
sjSDM(X = NULL, Y = NULL, formula = NULL, df = NULL,
  l1_coefs = 0, l2_coefs = 0, l1_cov = 0, l2_cov = 0, iter = 50L,
  step_size = NULL, learning_rate = 0.1, parallel = 0L, device = 2,
  dtype = "float32")
}
\arguments{
\item{X}{matrix of environmental predictors}

\item{Y}{matrix of species occurences/responses}

\item{formula}{formula object for predictors}

\item{df}{degree of freedom for covariance parametrization, if `NULL` df is set to `ncol(Y)/2`}

\item{l1_coefs}{strength of lasso regularization on environmental coefficients: `l1_coefs * sum(abs(coefs))`}

\item{l2_coefs}{strength of ridge regularization on environmental coefficients: `l2_coefs * sum(coefs^2)`}

\item{l1_cov}{strength of lasso regulIarization on covariances in species-species association matrix}

\item{l2_cov}{strength of ridge regularization on covariances in species-species association matrix}

\item{iter}{number of fitting iterations}

\item{step_size}{batch size for stochastic gradient descent, if `NULL` then step_size is set to: `step_size = 0.1*nrow(X)`}

\item{learning_rate}{learning rate for [optimizer_adamax]}
}
\description{
fast and accurate joint species model
}
\examples{
\donttest{

# simulate community:
com = simulate_SDM(env = 3L, species = 5L, sites = 100L)

# fit model:
model = sjSDM(X = com$env_weights, Y = com$response, iter = 10L)
coef(model)
summary(model)
getCov(model)

# fit model with interactions:
model = sjSDM(X = com$env_weights, Y = com$response, formula = ~X1:X2 + X3, iter = 10L)
summary(model)

# without intercept:
model = sjSDM(X = com$env_weights, Y = com$response, formula = ~ 0 + X1:X2 + X3, iter = 10L)
summary(model)

# predict with model:
preds = predict(model, newdata = com$env_weights)



}
}
