% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning.R
\name{sjSDM_cv}
\alias{sjSDM_cv}
\title{Cross validation of elastic net tuning}
\usage{
sjSDM_cv(Y, env = NULL, biotic = bioticStruct(), tune = c("random",
  "grid"), CV = 5L, tune_steps = 20L, alpha_cov = seq(0, 1, 0.1),
  alpha_coef = seq(0, 1, 0.1), lambda_cov = 2^seq(-10, -1, length.out = 20),
  lambda_coef = 2^seq(-10, -0.5, length.out = 20), n_cores = NULL, ...)
}
\arguments{
\item{Y}{species occurrence matrix}

\item{env}{matrix of environmental predictors or object of type \code{\link{envLinear}}, or \code{\link{envDNN}}}

\item{biotic}{defines biotic (species-species associations) structure, object of type \code{\link{bioticStruct}}. Alpha and lambda have no influence}

\item{tune}{tuning strategy, random or grid search}

\item{CV}{n-fold cross validation}

\item{tune_steps}{number of tuning steps}

\item{alpha_cov}{weighting of l1 and l2 on covariances: \eqn{(1 - \alpha) * |cov| + \alpha ||w||^2}}

\item{alpha_coef}{weighting of l1 and l2 on coefficients: \eqn{(1 - \alpha) * |coef| + \alpha ||coef||^2}}

\item{lambda_cov}{overall regularization strength on covariances}

\item{lambda_coef}{overall regularization strength on coefficients}

\item{n_cores}{number of cores for parallelization}

\item{...}{arguments passed to sjSDM, see \code{\link{sjSDM}}}
}
\description{
Cross validation of elastic net tuning
}
\examples{
\donttest{
# simulate sparse community:
com = simulate_SDM(env = 5L, species = 25L, sites = 100L, sparse = 0.5)

# tune regularization:
tune_results = sjSDM_cv(Y = com$response,
                        env = com$env_weights, 
                        tune = "random", # random steps in tune-paramter space
                        CV = 3L, # 3-fold cross validation
                        tune_steps = 25L,
                        alpha_cov = seq(0, 1, 0.1),
                        alpha_coef = seq(0, 1, 0.1),
                        lambda_cov = seq(0, 0.1, 0.001), 
                        lambda_coef = seq(0, 0.1, 0.001),
                        n_cores = 2L, # small models can be also run in parallel on the GPU
                        iter = 2L # we can pass arguments to sjSDM via ...
                        )

# print overall results:
tune_results

# summary (mean values over CV for each tuning step)
summary(tune_results)

# visualize tuning and best points:
best = plot(tune_results, perf = "logLik")

# fit model with new regularization paramter:
model = sjSDM(Y = com$response,
              env = envLinear(com$env_weights, 
                              lambda = best[["lambda_coef"]],
                              alpha = best[["alpha_coef"]]),
              biotic = bioticStruct(lambda = best[["lambda_cov"]],
                                    alpha = best[["alpha_cov"]])
              )

summary(model)
}
}
\seealso{
\code{\link{plot.sjSDM_cv}}, \code{\link{print.sjSDM_cv}}, \code{\link{summary.sjSDM_cv}}
}
