% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjSDM_configs.R
\name{sjSDMControl}
\alias{sjSDMControl}
\title{sjSDM control object}
\usage{
sjSDMControl(optimizer = Adamax(), scheduler = FALSE)
}
\arguments{
\item{optimizer}{object of type}

\item{scheduler}{use scheduler or not}
}
\description{
sjSDM control object
}
