% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjSDM_configs.R
\name{sjSDMControl}
\alias{sjSDMControl}
\title{sjSDM control object}
\usage{
sjSDMControl(optimizer = RMSprop(), scheduler = FALSE)
}
\arguments{
\item{optimizer}{object of type \code{\link{RMSprop}}, \code{\link{Adamax}}, \code{\link{SGD}}, \code{\link{AccSGD}}, or \code{\link{AdaBound}}}

\item{scheduler}{use scheduler or not}
}
\description{
sjSDM control object
}
