% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjSDM.R
\name{summary.sjSDM}
\alias{summary.sjSDM}
\title{Return summary of a fitted sjSDM model}
\usage{
\method{summary}{sjSDM}(object, ...)
}
\arguments{
\item{object}{a model fitted by \code{\link{sjSDM}}}

\item{...}{optional arguments for compatibility with the generic function, no functionality implemented}
}
\value{
The above matrix is silently returned.
}
\description{
Return summary of a fitted sjSDM model
}
