% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjSDM_cv.R
\name{summary.sjSDM_cv}
\alias{summary.sjSDM_cv}
\title{Return summary of a fitted sjSDM_cv model}
\usage{
\method{summary}{sjSDM_cv}(object, ...)
}
\arguments{
\item{object}{a model fitted by \code{\link{sjSDM_cv}}}

\item{...}{optional arguments for compatibility with the generic function, no functionality implemented}
}
\value{
Above data frame is silently returned.
}
\description{
Return summary of a fitted sjSDM_cv model
}
