% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjSDM.R
\name{update.sjSDM}
\alias{update.sjSDM}
\title{Update and re-fit a model call}
\usage{
\method{update}{sjSDM}(object, env_formula = NULL, spatial_formula = NULL, biotic = NULL, ...)
}
\arguments{
\item{object}{of class 'sjSDM'}

\item{env_formula}{new environmental formula}

\item{spatial_formula}{new spatial formula}

\item{biotic}{new biotic config}

\item{...}{additional arguments}
}
\value{
An S3 class of type 'sjSDM'. See \code{\link{sjSDM}} for more information.
}
\description{
Update and re-fit a model call
}
